+++ Changelog.md
@@ -12,6 +12,7 @@
 * Add parsing `fldSimple` inside hyperlinks
 * Add support of `wp14:pctPosHOffset` and `wp14:pctPosVOffset`
 * Add support of parsing `w:noBreakHyphen` in PargraphRun
+* Add support of Table Row Properties - Height
 
 ### Fixes
 * Fix parsing shape in paragraph run